Moving window and rendering to new function
- src/window.rs now contains the logic for creating and managing the
  window for displaying sensor information. src/main.rs is now a very
  thin interface.
- Added some docstrings.
